Move parametric tests into unit test module
