--- py-polars/tests/parametric/conftest.py
@@ -1,7 +1,0 @@
-import os
-
-from polars.testing.parametric import load_profile
-
-load_profile(
-    profile=os.environ.get("POLARS_HYPOTHESIS_PROFILE", "fast"),  # type: ignore[arg-type]
-)
