+++ py-polars/tests/unit/functions/test_business_day_count.py
@@ -1,8 +1,15 @@
+from __future__ import annotations
+
+import datetime as dt
 from datetime import date
 
+import hypothesis.strategies as st
+import numpy as np
 import pytest
+from hypothesis import assume, given, reject
 
 import polars as pl
+from polars._utils.various import parse_version
 from polars.testing import assert_series_equal
 
 
@@ -120,3 +127,37 @@ def test_business_day_count_w_holidays() -> None:
     )["business_day_count"]
     expected = pl.Series("business_day_count", [0, 5, 5], pl.Int32)
     assert_series_equal(result, expected)
+
+
+@given(
+    start=st.dates(min_value=dt.date(1969, 1, 1), max_value=dt.date(1970, 12, 31)),
+    end=st.dates(min_value=dt.date(1969, 1, 1), max_value=dt.date(1970, 12, 31)),
+    week_mask=st.lists(
+        st.sampled_from([True, False]),
+        min_size=7,
+        max_size=7,
+    ),
+    holidays=st.lists(
+        st.dates(min_value=dt.date(1969, 1, 1), max_value=dt.date(1970, 12, 31)),
+        min_size=0,
+        max_size=100,
+    ),
+)
+def test_against_np_busday_count(
+    start: dt.date, end: dt.date, week_mask: tuple[bool, ...], holidays: list[dt.date]
+) -> None:
+    assume(any(week_mask))
+    result = (
+        pl.DataFrame({"start": [start], "end": [end]})
+        .select(
+            n=pl.business_day_count(
+                "start", "end", week_mask=week_mask, holidays=holidays
+            )
+        )["n"]
+        .item()
+    )
+    expected = np.busday_count(start, end, weekmask=week_mask, holidays=holidays)
+    if start > end and parse_version(np.__version__) < parse_version("1.25"):
+        # Bug in old versions of numpy
+        reject()
+    assert result == expected
