+++ py-polars/tests/unit/functions/test_lit.py
@@ -5,9 +5,11 @@ from typing import Any
 
 import numpy as np
 import pytest
+from hypothesis import given
 
 import polars as pl
 from polars.testing import assert_frame_equal
+from polars.testing.parametric.strategies.data import datetimes
 
 
 @pytest.mark.parametrize(
@@ -95,3 +97,24 @@ def test_lit_unsupported_type() -> None:
         match="cannot create expression literal for value of type LazyFrame: ",
     ):
         pl.lit(pl.LazyFrame({"a": [1, 2, 3]}))
+
+
+@given(value=datetimes("ns"))
+def test_datetime_ns(value: datetime) -> None:
+    result = pl.select(pl.lit(value, dtype=pl.Datetime("ns")))["literal"][0]
+    assert result == value
+
+
+@given(value=datetimes("us"))
+def test_datetime_us(value: datetime) -> None:
+    result = pl.select(pl.lit(value, dtype=pl.Datetime("us")))["literal"][0]
+    assert result == value
+    result = pl.select(pl.lit(value, dtype=pl.Datetime))["literal"][0]
+    assert result == value
+
+
+@given(value=datetimes("ms"))
+def test_datetime_ms(value: datetime) -> None:
+    result = pl.select(pl.lit(value, dtype=pl.Datetime("ms")))["literal"][0]
+    expected_microsecond = value.microsecond // 1000 * 1000
+    assert result == value.replace(microsecond=expected_microsecond)
