+++ py-polars/tests/unit/operations/rolling/test_rolling.py
@@ -3,15 +3,22 @@ from __future__ import annotations
 from datetime import date, datetime, timedelta
 from typing import TYPE_CHECKING
 
+import hypothesis.strategies as st
 import numpy as np
 import pytest
+from hypothesis import assume, given
 from numpy import nan
 
 import polars as pl
+from polars._utils.convert import parse_as_duration_string
 from polars.exceptions import ComputeError, InvalidOperationError
 from polars.testing import assert_frame_equal, assert_series_equal
+from polars.testing.parametric import column, dataframes
+from polars.testing.parametric.strategies.dtype import _time_units
 
 if TYPE_CHECKING:
+    from hypothesis.strategies import SearchStrategy
+
     from polars.type_aliases import ClosedInterval, PolarsDataType, TimeUnit
 
 
@@ -1012,3 +1019,154 @@ def test_temporal_windows_size_without_by_15977() -> None:
             match=r"Passing a str to `rolling_\*` is deprecated"
         ):
             df.select(pl.col("a").rolling_mean("3d"))
+
+
+def interval_defs() -> SearchStrategy[ClosedInterval]:
+    closed: list[ClosedInterval] = ["left", "right", "both", "none"]
+    return st.sampled_from(closed)
+
+
+@given(
+    period=st.timedeltas(
+        min_value=timedelta(microseconds=0), max_value=timedelta(days=1000)
+    ).map(parse_as_duration_string),
+    offset=st.timedeltas(
+        min_value=timedelta(days=-1000), max_value=timedelta(days=1000)
+    ).map(parse_as_duration_string),
+    closed=interval_defs(),
+    data=st.data(),
+    time_unit=_time_units(),
+)
+def test_rolling_parametric(
+    period: str,
+    offset: str,
+    closed: ClosedInterval,
+    data: st.DataObject,
+    time_unit: TimeUnit,
+) -> None:
+    assume(period != "")
+    dataframe = data.draw(
+        dataframes(
+            [
+                column(
+                    "ts",
+                    strategy=st.datetimes(
+                        min_value=datetime(2000, 1, 1),
+                        max_value=datetime(2001, 1, 1),
+                    ),
+                    dtype=pl.Datetime(time_unit),
+                ),
+                column(
+                    "value",
+                    strategy=st.integers(min_value=-100, max_value=100),
+                    dtype=pl.Int64,
+                ),
+            ],
+            min_size=1,
+        )
+    )
+    df = dataframe.sort("ts")
+    result = df.rolling("ts", period=period, offset=offset, closed=closed).agg(
+        pl.col("value")
+    )
+
+    expected_dict: dict[str, list[object]] = {"ts": [], "value": []}
+    for ts, _ in df.iter_rows():
+        window = df.filter(
+            pl.col("ts").is_between(
+                pl.lit(ts, dtype=pl.Datetime(time_unit)).dt.offset_by(offset),
+                pl.lit(ts, dtype=pl.Datetime(time_unit))
+                .dt.offset_by(offset)
+                .dt.offset_by(period),
+                closed=closed,
+            )
+        )
+        value = window["value"].to_list()
+        expected_dict["ts"].append(ts)
+        expected_dict["value"].append(value)
+    expected = pl.DataFrame(expected_dict).select(
+        pl.col("ts").cast(pl.Datetime(time_unit)),
+        pl.col("value").cast(pl.List(pl.Int64)),
+    )
+    assert_frame_equal(result, expected)
+
+
+@given(
+    window_size=st.timedeltas(
+        min_value=timedelta(microseconds=0), max_value=timedelta(days=2)
+    ).map(parse_as_duration_string),
+    closed=interval_defs(),
+    data=st.data(),
+    time_unit=_time_units(),
+    aggregation=st.sampled_from(
+        [
+            "min",
+            "max",
+            "mean",
+            "sum",
+            "std",
+            "var",
+            "median",
+        ]
+    ),
+)
+def test_rolling_aggs(
+    window_size: str,
+    closed: ClosedInterval,
+    data: st.DataObject,
+    time_unit: TimeUnit,
+    aggregation: str,
+) -> None:
+    assume(window_size != "")
+
+    # Testing logic can be faulty when window is more precise than time unit
+    # https://github.com/pola-rs/polars/issues/11754
+    assume(not (time_unit == "ms" and "us" in window_size))
+
+    dataframe = data.draw(
+        dataframes(
+            [
+                column(
+                    "ts",
+                    strategy=st.datetimes(
+                        min_value=datetime(2000, 1, 1),
+                        max_value=datetime(2001, 1, 1),
+                    ),
+                    dtype=pl.Datetime(time_unit),
+                ),
+                column(
+                    "value",
+                    strategy=st.integers(min_value=-100, max_value=100),
+                    dtype=pl.Int64,
+                ),
+            ],
+        )
+    )
+    df = dataframe.sort("ts")
+    func = f"rolling_{aggregation}_by"
+    result = df.with_columns(
+        getattr(pl.col("value"), func)("ts", window_size=window_size, closed=closed)
+    )
+
+    expected_dict: dict[str, list[object]] = {"ts": [], "value": []}
+    for ts, _ in df.iter_rows():
+        window = df.filter(
+            pl.col("ts").is_between(
+                pl.lit(ts, dtype=pl.Datetime(time_unit)).dt.offset_by(
+                    f"-{window_size}"
+                ),
+                pl.lit(ts, dtype=pl.Datetime(time_unit)),
+                closed=closed,
+            )
+        )
+        expected_dict["ts"].append(ts)
+        if window.is_empty():
+            expected_dict["value"].append(None)
+        else:
+            value = getattr(window["value"], aggregation)()
+            expected_dict["value"].append(value)
+    expected = pl.DataFrame(expected_dict).select(
+        pl.col("ts").cast(pl.Datetime(time_unit)),
+        pl.col("value").cast(result["value"].dtype),
+    )
+    assert_frame_equal(result, expected)
